open source tech company
this is another project where i have showcased my frontend designing skills majority of my work was with CSS, HTML, a little bit with bootstrap. This is a tech company where worldwide startups can list themselves among the startup community and can share their technology with the community. In other word this is a social media o startups across the globe.